ai-services/requirements.txt
flask==2.3.3
flask-cors==4.0.0
transformers==4.35.0
torch>=2.2.0
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.3.2
numpy>=1.24.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
